Tue, May 3, 2022, 4:26 PM +00:00

+++ readability/readability.c
@@ -28,7 +28,7 @@ int main(void)
     }
     else
     {
-    printf("Grade %f\n" , round(index));
+    printf("Grade %i\n" ,(int) round(index));
     }
 }
